# PERF: Series(pyarrow-backed).rank