--- a/pandas/core/arrays/base.py
+++ b/pandas/core/arrays/base.py
@@ -1576,8 +1576,6 @@ class ExtensionArray:
         if axis != 0:
             raise NotImplementedError
 
-        # TODO: we only have tests that get here with dt64 and td64
-        # TODO: all tests that get here use the defaults for all the kwds
         return rank(
             self,
             axis=axis,