- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.0.0.rst` file if fixing a bug or adding a new feature.

Perf improvement in `Series(pyarrow-backed).rank` by using `pyarrow.compute.rank`. All parameter combinations use pyarrow compute functions except for `method="average"` which falls back to `algos.rank`.

```
import pandas as pd 
import numpy as np

ser = pd.Series(np.random.randn(10**6), dtype="float64[pyarrow]")

%timeit ser.rank(method="first")

# 1.41 s ± 10.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)    <- main
# 148 ms ± 1.42 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  <- PR
```